+++ src/renderers/stackingOffset.ts
@@ -175,9 +175,12 @@ export class StackingOffsetRenderer extends RendererBase {
             }
         }
 
-        // Finally, annotations
+        // annotations
         if (this.options.showAnnotations) {
             this.annotateBoard(gridPoints);
         }
+
+        // button bar
+        this.placeButtonBar(gridPoints);
     }
 }