+++ src/renderers/stackingTiles.ts
@@ -126,9 +126,12 @@ export class StackingTilesRenderer extends RendererBase {
             }
         }
 
-        // Finally, annotations
+        // annotations
         if (this.options.showAnnotations) {
             this.annotateBoard(gridPoints);
         }
+
+        // button bar
+        this.placeButtonBar(gridPoints);
     }
 }